requirements.txt
matplotlib==3.4.3
numpy==1.21.1
thop==0.0.31.post2005241907
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
yacs==0.1.8
Pillow==8.3.1
opencv-python==4.5.3.56
imgviz==1.2.6